requirements.txt
Django
djangorestframework
python-keycloak
psycopg
python-dotenv
drf-spectacular
channels
channels_redis